pyproject.toml
[tool.poetry]
name = "3xuigen"
version = "0.1.0"
description = "API для управления 3xui через туннель"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.1"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
httpx = "^0.27.0"
paramiko = "^3.4.0"  # Для SSH подключения

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"